requirements.txt
Django==5.1.6
python-dotenv==1.0.1
python-telegram-bot==13.15